New_Project-CausalEdit/test_basic_functionality.py
#!/usr/bin/env python3
"""
CausalEditor 基础功能测试脚本
验证项目核心组件是否能正常工作
"""

import sys
import logging
from pathlib import Path

# 添加项目路径
sys.path.append(str(Path(__file__).parent))

def test_imports():
    """测试核心模块导入"""
    print("🔍 测试模块导入...")
    
    try:
        # 测试核心组件导入
        from causal_editor import CausalEditor, CausalConflictDetector, CounterfactualEditor, VectorDatabase
        print("✅ 核心组件导入成功")
        
        # 测试预计算组件导入
        from causal_editor.precompute.knowledge_extractor import WikidataExtractor, CSVKnowledgeExtractor
        from causal_editor.precompute.fingerprint_builder import ActivationFingerprintBuilder
        from causal_editor.precompute.precompute_pipeline import PrecomputePipeline
        print("✅ 预计算组件导入成功")
        
        # 测试模型集成导入（可能存在兼容性问题，暂时跳过）
        try:
            from modeling_llms.modeling_llama_causal import CausalLlamaForCausalLM
            print("✅ 模型集成组件导入成功")
        except Exception as e:
            print(f"⚠️  模型集成组件导入失败（版本兼容性问题）: {str(e)[:100]}...")
            print("   建议: 检查PyTorch和transformers版本兼容性")
        
        return True
        
    except ImportError as e:
        print(f"❌ 导入失败: {e}")
        return False

def test_basic_dependencies():
    """测试基础依赖"""
    print("\n🔍 测试基础依赖...")
    
    try:
        import torch
        print(f"✅ PyTorch {torch.__version__}")
        
        import transformers
        print(f"✅ Transformers {transformers.__version__}")
        
        try:
            import faiss
            print(f"✅ FAISS {faiss.__version__}")
        except ImportError:
            print("⚠️  FAISS未安装，建议安装: conda install faiss-cpu -c conda-forge")
        
        import numpy as np
        print(f"✅ NumPy {np.__version__}")
        
        import pandas as pd
        print(f"✅ Pandas {pd.__version__}")
        
        import requests
        print("✅ Requests")
        
        import tqdm
        print("✅ tqdm")
        
        return True
        
    except ImportError as e:
        print(f"❌ 依赖缺失: {e}")
        return False

def test_configuration():
    """测试配置文件"""
    print("\n🔍 测试配置文件...")
    
    try:
        import json
        config_path = Path("configs/causal_editor_config.json")
        
        if config_path.exists():
            with open(config_path, 'r', encoding='utf-8') as f:
                config = json.load(f)
            print("✅ 配置文件加载成功")
            print(f"   模型: {config.get('model', {}).get('name', 'N/A')}")
            print(f"   编辑强度: {config.get('causal_editor', {}).get('edit_strength', 'N/A')}")
            return True
        else:
            print("⚠️  配置文件不存在")
            return False
            
    except Exception as e:
        print(f"❌ 配置文件错误: {e}")
        return False

def test_cuda_availability():
    """测试CUDA可用性"""
    print("\n🔍 测试CUDA可用性...")
    
    try:
        import torch
        if torch.cuda.is_available():
            device_count = torch.cuda.device_count()
            current_device = torch.cuda.current_device()
            device_name = torch.cuda.get_device_name(current_device)
            print(f"✅ CUDA可用，设备数量: {device_count}")
            print(f"   当前设备: {current_device} ({device_name})")
            return True
        else:
            print("⚠️  CUDA不可用，将使用CPU")
            return False
            
    except Exception as e:
        print(f"❌ CUDA检查失败: {e}")
        return False

def create_simple_test_data():
    """创建简单的测试数据"""
    print("\n🔍 创建测试数据...")
    
    try:
        # 创建简单的知识三元组测试数据
        test_triplets = [
            {
                "subject": "Einstein",
                "relation": "published_paper_on",
                "object": "Photoelectric Effect",
                "text": "Einstein published a paper on the photoelectric effect.",
                "confidence": 1.0,
                "source": "test"
            },
            {
                "subject": "Einstein",
                "relation": "received_award",
                "object": "Nobel Prize",
                "text": "Einstein received the Nobel Prize.",
                "confidence": 1.0,
                "source": "test"
            },
            {
                "subject": "Paris",
                "relation": "is_capital_of",
                "object": "France",
                "text": "Paris is the capital of France.",
                "confidence": 1.0,
                "source": "test"
            }
        ]
        
        # 保存测试数据
        import json
        test_data_path = Path("test_data")
        test_data_path.mkdir(exist_ok=True)
        
        with open(test_data_path / "test_triplets.json", 'w', encoding='utf-8') as f:
            json.dump(test_triplets, f, indent=2, ensure_ascii=False)
        
        print(f"✅ 测试数据已创建: {len(test_triplets)} 个三元组")
        return True
        
    except Exception as e:
        print(f"❌ 创建测试数据失败: {e}")
        return False

def main():
    """主测试函数"""
    print("🚀 CausalEditor 基础功能测试")
    print("=" * 50)
    
    # 设置日志
    logging.basicConfig(level=logging.INFO)
    
    test_results = []
    
    # 运行各项测试
    test_results.append(("模块导入", test_imports()))
    test_results.append(("基础依赖", test_basic_dependencies()))
    test_results.append(("配置文件", test_configuration()))
    test_results.append(("CUDA可用性", test_cuda_availability()))
    test_results.append(("测试数据", create_simple_test_data()))
    
    # 总结测试结果
    print("\n" + "=" * 50)
    print("📊 测试结果总结:")
    
    passed_count = 0
    for test_name, result in test_results:
        status = "✅ 通过" if result else "❌ 失败"
        print(f"   {test_name}: {status}")
        if result:
            passed_count += 1
    
    print(f"\n通过率: {passed_count}/{len(test_results)} ({passed_count/len(test_results)*100:.1f}%)")
    
    if passed_count >= len(test_results) - 1:  # 允许一个测试失败
        print("\n🎉 基础功能测试基本通过！")
        print("💡 建议:")
        print("   1. 运行预计算脚本构建知识库")
        print("   2. 在TruthfulQA数据集上测试效果")
        print("   3. 调整编辑强度等超参数")
    else:
        print("\n⚠️  存在多个问题，请检查依赖安装")
        print("💡 建议:")
        print("   1. 确保所有依赖包已正确安装")
        print("   2. 检查Python环境配置")
        print("   3. 参考README中的安装指南")

if __name__ == "__main__":
    main() 