New_Project-CausalEdit/requirements.txt
# CausalEditor 项目依赖
# 深度学习框架
torch>=2.0.0
transformers>=4.30.0

# 向量搜索和数据处理
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0

# 网络请求和进度显示
requests>=2.25.0
tqdm>=4.60.0

# 科学计算和可视化
scikit-learn>=1.0.0
matplotlib>=3.5.0

# 文本处理和工具
spacy>=3.4.0
regex>=2022.1.18

# 数据格式支持
jsonlines>=3.0.0
datasets>=2.0.0

# 开发和测试工具
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# 可选：GPU加速FAISS (如果有NVIDIA GPU)
# faiss-gpu>=1.7.0  # 替换faiss-cpu

# 可选：加速tokenizer
# tokenizers>=0.13.0 